fix: rename `componetns` package to `components`
