+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/screen/cropImage/components/UploadImageBox.kt
@@ -1,4 +1,4 @@
-package net.thechance.mena.dukan.presentation.screen.cropImage.componetns
+package net.thechance.mena.dukan.presentation.screen.cropImage.components
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
